requirements.txt
appdirs==1.4.4
bidict==0.21.2
black==20.8b1
click==7.1.2
dataclasses==0.8
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
regex==2020.11.13
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
